* Updated dependencies to latest Kotlin * Small fixes in DependencyResolverTest.kt

+++ src/test/kotlin/kscript/app/resolver/DependencyResolverTest.kt
@@ -2,6 +2,7 @@ package kscript.app.resolver
 
 import assertk.assertThat
 import assertk.assertions.contains
+import assertk.assertions.exists
 import assertk.assertions.isFailure
 import assertk.assertions.isInstanceOf
 import kscript.app.model.Dependency
@@ -12,6 +13,7 @@ import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
 import java.nio.file.Path
 import java.nio.file.Paths
+import kotlin.io.path.exists
 import kotlin.io.path.invariantSeparatorsPathString
 
 class DependencyResolverTest {
@@ -30,7 +32,9 @@ class DependencyResolverTest {
     @ValueSource(strings = ["log4j:log4j:1.2.14", "net.clearvolume:cleargl:jar:2.0.1"])
     fun `Resolve dependencies`(dependencyString: String) {
         val dependency = Dependency(dependencyString)
-        assertThat(dependencyResolver.resolve(setOf(dependency))).contains(calculateArtifactPath(dependency, true))
+        val calculatedPath = calculateArtifactPath(dependency, true)
+        assertThat(dependencyResolver.resolve(setOf(dependency))).contains(calculatedPath)
+        assertThat(calculatedPath).exists()
     }
 
     @Test
@@ -54,7 +58,9 @@ class DependencyResolverTest {
 
         if (cleanupFirst) {
             val cleanupPath = Paths.get("$repositoryPathString/$group")
-            FileUtils.cleanDirectory(cleanupPath.toFile())
+            if (cleanupPath.exists()) {
+                FileUtils.cleanDirectory(cleanupPath.toFile())
+            }
         }
 
         return calculatedPath
